Hold the Windows client components to the /W4 warning bar
The ACM codec and the DirectShow filter build again, but neither had ever
been measured against the warning level the rest of the tree is held to.
At /W4 they were not clean, and neither was the small sample encoder that
sits in the same solution.

Most of what the level reported is not this project's code: the vendored
XML parser and the DirectShow base classes, which the filter compiles
directly because it derives from the classes they implement.  Neither can
be acted on without forking it, so both are now compiled at their own
author's warning level and the bar measures what is left.

Of the rest, the unreferenced formal parameters are the case the library
already suppresses for this compiler - debug-output stubs and window
procedures that must accept a signature they do not use - so the clients
now share that policy instead of being the only place it does not apply.
The deliberate narrowing conversions are cast at the site, with what
bounds the value written beside them, rather than disabled wholesale.

The remainder was ordinary: an assignment used as a condition, three
locals that were never read, and two returns after a switch whose every
case already returns.

Two narrowing conversions in the ACM configuration writer are left alone.
They put a fractional ratio into an integer attribute, so what they change
is a stored value and not only its type; that is a question about the
configuration format and is handled on its own.

+++ lame/ACM/ACM.cpp
@@ -1226,14 +1226,18 @@ void ACM::GetMP3FormatForIndex(const DWORD the_Index, WAVEFORMATEX & the_Format,
 		else
 			Block_size = 576;
 	
-		the_Format.nChannels = bitrate_table[the_Index].channels;
+		// Both narrowings are into WAVEFORMATEX's 16-bit fields and both are
+		// bounded well inside it: the channel count is 1 or 2, and the block
+		// size is the byte length of one frame - at most 1152 samples at 320
+		// kbit/s and 32 kHz, under 1500 bytes.
+		the_Format.nChannels = (WORD) bitrate_table[the_Index].channels;
 
 		the_Format.cbSize = sizeof(MPEGLAYER3WAVEFORMAT) - sizeof(WAVEFORMATEX);
 		MPEGLAYER3WAVEFORMAT * tmpFormat = (MPEGLAYER3WAVEFORMAT *) &the_Format;
 		tmpFormat->wID             = 1;
 		// this is the only way I found to know if we do CBR or ABR
 		tmpFormat->fdwFlags        = 2 + ((bitrate_table[the_Index].mode == vbr_abr)?0:2);
-		tmpFormat->nBlockSize      = Block_size * the_Format.nAvgBytesPerSec / the_Format.nSamplesPerSec;
+		tmpFormat->nBlockSize      = (WORD) (Block_size * the_Format.nAvgBytesPerSec / the_Format.nSamplesPerSec);
 		tmpFormat->nFramesPerBlock = 1;
 		tmpFormat->nCodecDelay     = 0; // 0x0571 on FHG
